added validation trait
